Add jira-cli
This addition brought forward some issues with the go builder. The
problem is that goreleaser files can vary greatly and sometimes use
feature that my workflow can't easily workaround. I ran into this
issue the other day when updating fzf too. So I decided to just make
a custom barebones .goreleaser.yml to use for the builds. It is unclear
to me yet whether this will cause issues long term.